Created classes for TrelloAPIClient and Item.
TrelloAPIClient handles calls to Trello API for retrieving, updating creating and deleting cards. Item class holds the card specific data.

Integrated the API calls to replace the existing app functionality for retrieving, adding, deleting and updating to-do items.

Added a route for a 'doing' status.

Added form elements for adding a due date and description to the to-do items.

Added 'card' bootstrap styling to the to-do item.